+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added Platine Gestion redirection
+- Add a button to leave the questionnaire and go back to the base app
 
 ### Changed